idam_uam_backend-master/requirements.txt
amqp==2.3.2
astroid==2.0.4
autopep8==1.4.4
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
defusedxml==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.0.0
django-polymorphic==2.0.3
django-pyodbc-azure==2.1.0.0
django-rest-polymorphic==0.1.8
django-spa==0.2.0
djangorestframework==3.9.0
dnspython==1.15.0
eventlet==0.24.1
factory-boy==2.11.1
Faker==0.9.2
greenlet==0.4.15
idna==2.7
importlib-metadata==0.6
ipython==7.5.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.4
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
kombu==4.2.1
lazy-object-proxy==1.3.1
lxml==4.3.4
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
monotonic==1.5
nameko==2.12.0
parso==0.4.0
path.py==11.5.0
pbr==5.1.1
pickleshare==0.7.5
pkgconfig==1.5.1
prompt-toolkit==2.0.9
pycodestyle==2.5.0
Pygments==2.4.1
pylint==2.1.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pyodbc==4.0.24
python-dateutil==2.7.5
python3-saml==1.6.0
pytz==2018.7
PyYAML==5.1.2
requests==2.20.0
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==3.2.2
wrapt==1.10.11
#xmlsec==1.3.52.dev0
